service/Services/entities/src/main/java/org/infy/entities/triggerinputs/BuildDeployEnv.java
/***********************************************************************************************
*
* Copyright 2018 Infosys Ltd.
* Use of this source code is governed by MIT license that can be found in the LICENSE file or at
* https://opensource.org/licenses/MIT.
*
***********************************************************************************************/

package org.infy.entities.triggerinputs;

import java.util.List;

import com.google.gson.annotations.Expose;
import com.google.gson.annotations.SerializedName;

/**
 * Entity to build and deploy environment detail
 * 
 * @author Infosys
 *
 */
public class BuildDeployEnv {

	@SerializedName("releaseNumber")
	@Expose
	private String releaseNumber;
	@SerializedName("Env")
	@Expose
	private List<String> env = null;

	public String getReleaseNumber() {
		return releaseNumber;
	}

	public void setReleaseNumber(String releaseNumber) {
		this.releaseNumber = releaseNumber;
	}

	public List<String> getEnv() {
		return env;
	}

	public void setEnv(List<String> env) {
		this.env = env;
	}

}
